Implemented correct "real focal length".
In the Hugin models (Poly3 and PTLens), the focal length is incorrect because
the first derivative of the distortion polynomial is not unity.  Therefore, the
focal length has to be corrected by that factor.  This change is a no-op,
because the factor is taken out again when the real focal length is applied in
the projection transformation.  However, it is the first step towards the
<field_of_view> tag becoming obsolete.