Switched to Hugin-compatible interpretation of nominal focal length.
If no <real-focal-length> tag is present, Lensfun uses the nominal focal length
as a default.  I changed the code so that it interprets the nominal focal
length in terms of Hugin's quirky definition.  This way, distortion data
yielded by Hugin can be used in Lensfun out-of-the-box, even if the real focal
length was not measured.  Note that all of this is only significant for
non-rectilinear lenses.