Improved scoring of mount matching when doing a lens search
Formerly, if two <lens> entries differed only in *additional* <mount> entries
in one of the two, this one got lower scoring, although the required mount was
available in both entries.  Now, only the availability of the correct mount in
the lens entry is scored, no matter how many others are there.