Enhanced vignetting parameter interpolation for ACM model.
The problem is that the Adobe models express coordinates in units of the focal
length.  Thus, for longer focal lengths, the coordinates become numerically
small.  However, the lens errors usually don't scale equally.  The result is
that the parameters become large to huge for longer focal lengths.  This should
not be a problem for TCA and distortion (we'll see) but it is for the rather
fragile IDW interpolation of vignetting parameters.  Therefore, I scale the
parameters down as if the coordinate system was in millimeters, interpolate,
and scale back again.