Improved vignetting interpolation.
Now, interpolation fails not so often anymore.  It used to be too strict: The
"distance" between f-stop numbers 1.4 and 2 is already 1.7.  If the focal
length doesn't match perfectly either, this may easily exceed 2, and any point
inbetween is farther than 1 away from a calibration point, so no correction is
done.