Stored focal length in a new attribute in lfModifier.
This seems to be ugly because it is the only EXIF parameter stored in
lfModifier.  It is needed by perspective correction.  I don't want to pass it
in Initialize_Perspective_Correction because on the long run, all EXIF data
should be passed to lfModifier's constructor and stored in the object.