Got rid of get_hugin_focal_correction.
Now, the coordinates refer to images of the correct focal length (up to scaling
of course).  In other words, the centre of the image is not scaled anymore by
un-distortion.  This happened for the PT-derived models PTLens and poly3
because of that unfortunate "d" parameter.

For geometry transform and -- possibly in the future -- perspective correction,
the correct focal length is important.  One can use get_hugin_focal_correction
for this, but I think a chain of focal-length-preserving image operators is
easier to understand, to maintain, and to extend.

Note that this change does not come with any performance or accuracy penalty.