Fixed width and height of the image.
Width and height are normally measured at the rim of the image.  However, the
grid of the pixel centres is one pixel smaller in each dimension.  And we
transform the coordinates of the pixel centres after all.

This bug caused wrong CenterX/CenterY values and wrong calculation of the
aspect ratio.  Of course, it only matters really for extremly small images.
But it also made creating test suites more difficult due to surprising
behaviour in the floating point results.