Corrected optical axis in case of non-maching sensor sizes.
This basically reverts commit 77a9bd4e.  I suspect I introduced 77a9bd4e in
order to fix something that was actually a fencing error.  But be that as it
may, the optical axis is a property of the lens rather than the sensor, so
lens->CenterX/Y must refer to normalized *lens* coordinates instead of
normalized sensor coordinates.

Now, all tests succeed, in particular the one added in commit 2100088.