Added LX10/LX15 upload 004463 #306, could be better, though.
These images exhibited some strange differences in distortion between the top and bottom of the frame in the wide shots, although this could have been due to the lower part of the building being very close. Correction seemed ok on an image from a different upload (not really sutable for calibration).