Added Powershot SX60 HS upload 093228 #393
There seems to be a large amount of sample variation in terms of offset with this "ultrazoom" camera. These values were derived by allowing Hugin to optimize d and e parameters, which produced good correction on the specific images used, then discarding the offset for the profile. The quality of correction will therefore vary depending on the eccentricity of the particular copy. Also, TCA correction became impossible (but also less needed) towards the longer end of the zoom, so hand-adjusted values were included only where they seemed to do more good than harm.