Add vignetting correction for Nikkor Z 85mm f/1.8 S
This adds a new section to the lens database and for
the moment just vignetting calibration data for a Nikkor Z 85mm f/1.8 S.

Shots were taken on this cloudy morning with a Nikon Z6.
I have verified the applied correction using darktable git master (3.1.0+2411)
and results look promising to me.

Distortion and TCA will follow some day when I'm able to find
suitable conditions according to the calibration processes.

Signed-off-by: René Kowalke <rene.kowalke@gmail.com>