Fix out-of-bound memory write in `lf_mlstr_add` (#1859)
* Fixed memory corruption in lf_mlstr_add
The memory corruption occurs when replacing the default string
with a new value that is larger. The code was written
assuming the new value was shorter; when that is not the case, an
out-of-bound write occurs. In the case I encountered, this corrupted
the allocation metadata, and caused realloc to fail with the cryptic
message "corrupted size vs. prev_size", then segfault. The old code was
also relying on undefined behavior when the input `str` variable was
NULL and `lang` was also NULL. In this case, `memcpy` was called on an
invalid address. However, because this requested a copy of zero
elements, the implementation was able to just let it slide and not
report an error. But this is implementation-defined, and should not be
relied upon.

* Added test for lf_mlstr_add
Prevents regression of 407a8606abc78b368b109a3b68f0d35a3d3c909e

* Use memmove instead of memcpy for overlapping ranges