Fix Meike 28mm f/2.8 lens name (#1964)
"MK-F" doesn't seem to stand for a known designation. If it were a part of identification string in Exif, it would be necessary to keep it for automatic lens ID matching. (Human readable name without MK-F could still be had in `<model lang=` line.) In this case however removing MK-F should be harmless.