Add M39 mount, next to M39/1
There are variants of the true Leica Thread-Mount (LTM) which is  is 39 mm in diameter and has a thread of 26 turns-per-inch or threads-per-inch (tpi) (approximately 0.977 mm pitch) of Whitworth thread form.

The Soviets in the 1930s produced their early FED cameras in M39×1 (39 mm by 1 mm DIN thread).

In the lensfun database some lenses are listed as M39 and some as M39/1.

Therefor, make sure both M39/1 and M39 are in generic.xml